rimossa libft, aggiunti file con le funzioni, splittate e fatta una minipulizia norm
