+++ includes/minishell.h
@@ -6,8 +6,10 @@
 # include <stdio.h>
 # include <readline/readline.h>
 # include <readline/history.h>
-
+# include <signal.h>
 # include <stdbool.h>
+# include <fcntl.h>
+# include <termios.h>
 
 # ifdef __APPLE__
 #  define rl_clear_history clear_history
@@ -16,7 +18,7 @@
 # define SUCCESS 0
 # define FAILURE 1
 
-# include "../libraries/ft_libft/libft.h"
+extern volatile sig_atomic_t g_signal;
 
 
 typedef enum	e_token_type
@@ -63,7 +65,7 @@ typedef struct s_pipeline
 }	t_pipeline;
 
 
-
+int		is_prev_not_redirection(t_token *prev);
 int		ft_strcmp(const char *main, const char *compared);
 int		ft_isspace(char c);
 int		has_closed_quotes(const char *str);
@@ -71,7 +73,6 @@ int		execute_command(t_cmd *cmd, char **envp);
 int		handle_command(t_cmd *cmd, char ***envp_new, int last_exit_status, int *should_exit, t_token *token);
 int		is_builtin(char *cmd);
 int		execute_builtin(char **argv, char ***envp, int exit_status, int *should_exit, t_token *token);
-void	hello_minishell(void);
 void	add_token(t_token **tokens, t_token *new_token);
 void	tokkenize_arrows(char *input, int *i, t_token **tokens);
 void	tokkenize_pipe(int *index, t_token **tokens);
@@ -105,6 +106,9 @@ void	free_envp_old(char ***envp_old, char **envp_new);
 int		builtin_pwd(char **argv);
 int		builtin_unset(char **argv, char ***envp_new);
 int		builtin_exit(char **argv, int exit_status, int *should_exit);
+int		is_redirection(t_token_type type);
+int		is_prev_not_redirection(t_token *prev);
+int		fill_command_data(t_cmd *cmd, t_token *start, t_token *end);
 char	*join_export_args(char **argv);
 
 
@@ -114,7 +118,26 @@ t_cmd	*create_cmd(void);
 void	free_pipeline(t_pipeline *pipeline);
 void	free_command(t_cmd *cmd);
 
+//libft
+int		ft_atoi(const char *nptr);
+int		ft_isalnum(int c);
+char	*ft_strjoin(char const *s1, char const *s2);
+void	ft_putstr_fd(char *s, int fd);
+size_t	ft_strspn(const char *s, const char *accept);
+void	*ft_memset(void *s, int c, size_t len);
+char	*ft_strdup(const char *s);
+size_t	ft_strlen(const char *s);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
+char	*ft_substr(char const *s, unsigned int start, size_t len);
+char	*ft_itoa(int n);
+char	**ft_split(char const *s, char c);
+char	*ft_strchr(const char *s, int c);
 
+//signals
+void	handle_sigint(int signum);
+void	handle_sigquit(int signum);
+void	setup_signals(void);
+void	disable_echoctl(void);
 
 
 #endif
