+++ srcs/05_builtin/export01.c
@@ -128,7 +128,7 @@ int	builtin_export(char **argv, char ***envp_new)
 	{
 		if (is_valid_key(argv[1]))
 		{
-			write(STDERR_FILENO, "Not a valid key\n", 16);
+			ft_putstr_fd("Not a valid key\n", STDERR_FILENO);
 			return (1);
 		}
 		if (key_exists(*envp_new, argv[1]) < 0)
